remove AbstractRepositoryConfig
